refactor: move named GraphQL query/mutation variables to own files
what was done (and nothing else):
 - moved GraphQL queries, that were in variables, to queries.py in order
   of their appearance in the original source file
 - moved GraphQL mutations, that were in variables, to mutations.py in
   order of their appearance in the original source file
 - imported the moved queries from queries.py
 - imported the moved mutations from mutations.py
 - updated snapshots, they were out-of-sync for some reason,
   the snapshots should not have changed because of this commit

NOTE:
 - variables to queries/mutations were NOT moved but left where they
   were in the original source file

refs KK-1143 (refactoring came up in code review discussion)
