+++ gdpr/tests/snapshots/snap_test_gdpr_api.py
@@ -681,6 +681,10 @@ Daniellechester, KS 04855""",
                                     "children": [],
                                     "key": "FREE_SPOT_NOTIFICATION_SUBSCRIPTIONS",
                                 },
+                                {
+                                    "key": "NOTES",
+                                    "value": "GDPR sensitive test notes for child",
+                                },
                             ],
                             "key": "CHILD",
                         }
